refactor+feat!!!: completely discard most of old code and add realms.
this commit adds a WIP realms feature, allowing a single server
to run multiple worlds that players can switch between.

a realm contains the world and all the players that reside within it.
every time a player switches between realms, they receive the new world
data and also all of the players that reside there, while the old players
just despawn.

this kinda breaks optimizations a ton, since it adds another layer of
linq, but for a wip commit i am very happy with it.