feat: compile plugins with Roslyn!!!
we now do a very dumb compilation pass on anything that's inside of
the "plugins/" directory.

if there's a .cs file inside of there, the plugin manager will pick
it up, check if it was already compiled, and if not, compile and load
it.

this allows us to pretty much enable C# scripting for Tako!

in the future i'd like it to be able to compile multiple files zipped
together, so plugins could be downloaded just as a repo archive, or
cloned directly into the plugins dir.